+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+
+## 0.1.1
+
+### Changed
+* The plugin name is now finalised to `LiviUtils`
+
+### Added
+* Integration tests :3
+
 ## 0.1.0
 
 initial release